+++ src/thirteenducks/cor/mainmenu/components/ScaledImage.java
@@ -53,7 +53,7 @@ public class ScaledImage extends Component {
      */
     Color color;
 
-    public ScaledImage(MainMenu mainMenuReference, int x, int y, double width, double height, String imagepath) {
+    public ScaledImage(MainMenu mainMenuReference, double x, double y, double width, double height, String imagepath) {
         super(mainMenuReference, x, y, width, height);
 
         imagePath = imagepath;
